feat: separate all room context responsabilities
